refactor(dive_3d): extract niceStep into nice_step.dart
